src/requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==8.1.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==2.2.2
gin-config==0.5.0
idna==2.10
importlib-metadata==5.2.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab==2.2.5
jupyterlab-server==1.2.0
lxml==4.9.2
MarkupSafe==2.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.24.1
packaging==20.4
pandas==1.5.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycparser==2.20
pydocstyle==6.1.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2022.7
pyzmq==19.0.2
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.2.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Unidecode==1.3.6
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.11.0
